wip 検索結果をフロント渡すところまで完了。ただし、検索結果一覧は親コンポーネントで出力しているため、親コンポーネントで定義するように修正が必要。proosで渡すならはじめからクライアント側で実行した方がいい？それか検索専用ページに飛ばすか
