+++ app/factory/[id]/page.tsx
@@ -1,7 +1,8 @@
 import Container from '@/app/_components/layouts/container'
 import Map from '@/app/_components/ui-element/googleMap'
 import Spinner from '@/app/_components/ui-parts/spinner'
-import { getAllFactory, getFactoryById } from '@/app/_features/factory/api'
+import { getAllFactory } from '@/app/_features/factory/api/getAllFactory'
+import { getFactoryById } from '@/app/_features/factory/api/getFactoryById'
 import { FavoriteButton } from '@/app/_features/favorite/components/favoriteButtton'
 import { getReviewsByFactoryId } from '@/app/_features/review/api'
 import {
