+++ app/page.tsx
@@ -1,5 +1,3 @@
-import { getAllFactory } from './_features/factory/api'
-
 export default async function Home() {
   return <main></main>
 }
